authorization , login and register model User

+++ alembic/versions/1cfc2b1d5469_venues_table.py
@@ -9,7 +9,7 @@ from typing import Sequence, Union
 
 from alembic import op
 import sqlalchemy as sa
-
+from sqlalchemy.engine.reflection import Inspector
 
 # revision identifiers, used by Alembic.
 revision: str = '1cfc2b1d5469'
@@ -19,13 +19,20 @@ depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
-    op.create_table(
-        'venues',
-        sa.Column('id', sa.String, primary_key=True, index=True),
-        sa.Column('name', sa.String, nullable=False, index=True),
-        sa.Column('url', sa.String),
-    )
+    bind = op.get_bind()
+    inspector = Inspector.from_engine(bind)
+    if 'venues' not in inspector.get_table_names():
+        op.create_table(
+            'venues',
+            sa.Column('id', sa.String, nullable=False, primary_key=True),
+            sa.Column('name', sa.String, nullable=False),
+            sa.Column('url', sa.String),
+        )
 
 
-def downgrade() -> None:
-    op.drop_table('venues')
+def downgrade():
+    # Drop the table if it exists
+    bind = op.get_bind()
+    inspector = Inspector.from_engine(bind)
+    if 'venues' in inspector.get_table_names():
+        op.drop_table('venues')
